requirements.txt
appdirs==1.4.4
attrs==21.2.0
autopage==0.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cliff==3.10.0
cmd2==2.3.2
colorama==0.4.4
cors==1.0.1
cryptography==36.0.0
debtcollector==2.3.0
decorator==5.1.0
dogpile.cache==1.1.4
filelock==3.4.0
Flask==2.0.2
Flask-Cors==3.0.10
future==0.18.2
gevent==21.8.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.2
iso8601==1.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.2
keystoneauth1==4.4.0
MarkupSafe==2.0.1
msgpack==1.0.3
munch==2.5.0
netaddr==0.8.0
netifaces==0.11.0
openstacksdk==0.60.0
os-service-types==1.7.0
osc-lib==2.4.2
oslo.config==8.7.1
oslo.i18n==5.1.0
oslo.serialization==4.2.0
oslo.utils==4.12.0
packaging==21.3
pbr==5.8.0
prettytable==2.4.0
pycodestyle==2.8.0
pycparser==2.21
pyparsing==3.0.6
pyperclip==1.8.2
pyreadline==2.1
PySocks==1.7.1
python-cinderclient==8.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-keystoneclient==4.3.0
python-novaclient==17.6.0
python-openstackclient==5.7.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-file==1.5.1
requestsexceptions==1.4.0
rfc3986==1.5.0
simplejson==3.17.6
six==1.16.0
stevedore==3.5.0
tldextract==3.1.2
toml==0.10.2
typing-extensions==4.0.0
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
